# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Activo Fijo API — CLT S.A.
  version: "1.0.0"
  description: |
    API REST para el Sistema de Gestión de Activos Fijos de CLT S.A.
    Arquitectura Hexagonal · FastAPI · PostgreSQL 15
  contact:
    name: CLT — Equipo de Desarrollo
    email: desarrollo@clt.com.py
servers:
  - url: http://localhost:8000
    description: Desarrollo local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, example: 28800 }
        usuario:
          $ref: '#/components/schemas/UsuarioResponse'

    # ── Activo ────────────────────────────────────────────
    ActivoCreate:
      type: object
      required: [nombre, categoria_id, valor_adquisicion]
      properties:
        nombre: { type: string, maxLength: 300 }
        categoria_id: { type: string, format: uuid }
        marca: { type: string, maxLength: 100 }
        modelo: { type: string, maxLength: 100 }
        numero_serie: { type: string, maxLength: 150 }
        fecha_compra: { type: string, format: date }
        valor_adquisicion: { type: number, minimum: 0 }
        vida_util_años: { type: integer, minimum: 1, default: 5 }
        valor_residual: { type: number, minimum: 0, default: 0 }
        area: { type: string }
        centro_costo_id: { type: string, format: uuid }
        responsable: { type: string }
        ubicacion: { type: string }
        foto_url: { type: string }
    ActivoUpdate:
      type: object
      properties:
        nombre: { type: string }
        marca: { type: string }
        modelo: { type: string }
        numero_serie: { type: string }
        fecha_compra: { type: string, format: date }
        valor_adquisicion: { type: number, minimum: 0 }
        vida_util_años: { type: integer, minimum: 1 }
        valor_residual: { type: number, minimum: 0 }
        estado:
          type: string
          enum: [activo, en_mantenimiento, dado_de_baja, reservado]
        area: { type: string }
        centro_costo_id: { type: string, format: uuid }
        responsable: { type: string }
        ubicacion: { type: string }
        foto_url: { type: string }
    ActivoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        codigo: { type: string, example: ACT-0001 }
        nombre: { type: string }
        categoria_id: { type: string, format: uuid }
        categoria_nombre: { type: string }
        marca: { type: string }
        modelo: { type: string }
        numero_serie: { type: string }
        fecha_compra: { type: string, format: date }
        valor_adquisicion: { type: number }
        vida_util_años: { type: integer }
        valor_residual: { type: number }
        depreciacion_anual: { type: number }
        depreciacion_acumulada: { type: number }
        valor_libro_actual: { type: number }
        estado:
          type: string
          enum: [activo, en_mantenimiento, dado_de_baja, reservado]
        area: { type: string }
        centro_costo_id: { type: string, format: uuid }
        centro_costo_nombre: { type: string }
        responsable: { type: string }
        ubicacion: { type: string }
        foto_url: { type: string }
        documentos:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PaginatedActivos:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ActivoResponse' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        total_pages: { type: integer }
...
    CategoriaCreate:
      type: object
      required: [nombre]
      properties:
        nombre: { type: string }
        icono: { type: string, default: package }
        color_hex: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
        vida_util_default_años: { type: integer, minimum: 1, default: 5 }
    CategoriaResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        icono: { type: string }
        color_hex: { type: string }
        vida_util_default_años: { type: integer }
        total_activos: { type: integer }
        activos_activos: { type: integer }
        activos_baja: { type: integer }

    # ── CentroCosto ───────────────────────────────────────
...
    AsignacionCreate:
      type: object
      required: [activo_id, empleado_nombre, area, fecha_asignacion]
      properties:
        activo_id: { type: string, format: uuid }
        empleado_nombre: { type: string }
        empleado_cedula: { type: string }
        area: { type: string }
        centro_costo_id: { type: string, format: uuid }
        fecha_asignacion: { type: string, format: date }
        observaciones: { type: string }
    AsignacionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        activo_id: { type: string, format: uuid }
        activo_codigo: { type: string }
        activo_nombre: { type: string }
        empleado_nombre: { type: string }
        empleado_cedula: { type: string }
        area: { type: string }
        centro_costo_id: { type: string, format: uuid }
        centro_costo_nombre: { type: string }
        fecha_asignacion: { type: string, format: date }
        fecha_baja: { type: string, format: date }
        observaciones: { type: string }
        estado:
          type: string
          enum: [vigente, baja]
        created_at: { type: string, format: date-time }

    # ── Mantenimiento ─────────────────────────────────────
...
    DispositivoCreate:
      type: object
      required: [nombre, tipo, protocolo]
      properties:
        nombre: { type: string }
        tipo:
          type: string
          enum: [impresora, lector_qr, lector_barcode]
        protocolo:
          type: string
          enum: [usb, tcp_ip, bluetooth, usb_hid]
        driver: { type: string }
        ip_address: { type: string }
    DispositivoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        tipo: { type: string }
        protocolo: { type: string }
        driver: { type: string }
        ip_address: { type: string }
        estado: { type: string }
        ultima_conexion: { type: string, format: date-time }

    # ── Usuario ───────────────────────────────────────────
    UsuarioResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        nombre_completo: { type: string }
        rol: { type: string }
        activo: { type: boolean }

    # ── Errors ────────────────────────────────────────────
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login con email y password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        "200":
          description: Token JWT generado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        "401": { description: Credenciales inválidas }
  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Información del usuario autenticado
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioResponse' }
        "401": { description: No autenticado }

  # ── Activos ──────────────────────────────────────────────
  /v1/activos:
    get:
      tags: [Activos]
      summary: Listar activos con filtros y paginación
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
        - { name: q, in: query, schema: { type: string } }
        - { name: categoria_id, in: query, schema: { type: string, format: uuid } }
        - { name: estado, in: query, schema: { type: string } }
        - { name: centro_costo_id, in: query, schema: { type: string, format: uuid } }
        - { name: area, in: query, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedActivos' }
    post:
      tags: [Activos]
      summary: Registrar nuevo activo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivoCreate' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivoResponse' }
        "400": { description: Validación fallida }
        "409": { description: Serie duplicada }
...
  /v1/activos/depreciacion:
    get:
      tags: [Activos]
      summary: Tabla de depreciación de todos los activos
      parameters:
        - { name: categoria_id, in: query, schema: { type: string, format: uuid } }
        - { name: area, in: query, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array }
                  totales: { type: object }
  /v1/activos/{id}:
    get:
      tags: [Activos]
      summary: Detalle de activo con asignaciones y mantenimientos
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivoResponse' }
        "404": { description: Activo no encontrado }
    put:
      tags: [Activos]
      summary: Actualizar activo completo
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivoUpdate' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivoResponse' }
    delete:
      tags: [Activos]
      summary: Dar de baja activo (soft delete)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Activo dado de baja }
        "409": { description: Tiene asignación vigente activa }
  /v1/activos/{id}/qr:
    get:
      tags: [Activos]
      summary: Generar QR PNG del activo
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: format, in: query, schema: { type: string, enum: [png, zpl], default: png } }
      responses:
        "200":
          content:
            image/png:
              schema: { type: string, format: binary }
  /v1/activos/{id}/foto:
    post:
      tags: [Activos]
      summary: Subir foto del activo
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                foto: { type: string, format: binary }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  foto_url: { type: string }

  # ── Categorías ────────────────────────────────────────────
  /v1/categorias:
    get:
      tags: [Categorías]
      summary: Listar categorías
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CategoriaResponse' }
    post:
      tags: [Categorías]
      summary: Crear categoría
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoriaCreate' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoriaResponse' }
  /v1/categorias/{id}:
    put:
      tags: [Categorías]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoriaCreate' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoriaResponse' }

  # ── Centros de Costo ──────────────────────────────────────
...
  /v1/asignaciones:
    get:
      tags: [Asignaciones]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
        - { name: estado, in: query, schema: { type: string } }
        - { name: activo_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AsignacionResponse' }
                  total: { type: integer }
    post:
      tags: [Asignaciones]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AsignacionCreate' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AsignacionResponse' }
        "409": { description: Activo ya tiene asignación vigente }
  /v1/asignaciones/{id}/baja:
    patch:
      tags: [Asignaciones]
      summary: Dar de baja asignación (devolución)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AsignacionResponse' }

  # ── Mantenimientos ────────────────────────────────────────
...
  /v1/dispositivos:
    get:
      tags: [Dispositivos]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DispositivoResponse' }
    post:
      tags: [Dispositivos]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DispositivoCreate' }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DispositivoResponse' }
  /v1/dispositivos/{id}:
    put:
      tags: [Dispositivos]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DispositivoCreate' }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DispositivoResponse' }

  # ── Reportes ──────────────────────────────────────────────
...